+++ src/sphinx_codeautolink/extension/__init__.py
@@ -75,6 +75,7 @@ class SphinxCodeAutoLink:
         self.warn_missing_inventory = None
         self.warn_failed_resolve = None
         self.warn_no_backreference = None
+        self.highlight_lang = None
 
         # Populated once
         self.outdated_docs: set[str] = set()
@@ -105,6 +106,7 @@ class SphinxCodeAutoLink:
         self.warn_missing_inventory = app.config.codeautolink_warn_on_missing_inventory
         self.warn_failed_resolve = app.config.codeautolink_warn_on_failed_resolve
         self.warn_no_backreference = app.config.codeautolink_warn_on_no_backreference
+        self.highlight_lang = app.config.highlight_language
 
         # Append static resources path so references in setup() are valid
         app.config.html_static_path.append(
@@ -138,6 +140,7 @@ class SphinxCodeAutoLink:
             global_preface=self.global_preface,
             custom_blocks=self.custom_blocks,
             concat_default=self.concat_default,
+            default_highlight_lang=self.highlight_lang,
         )
         doctree.walkabout(visitor)
         self.cache.transforms[visitor.current_document] = visitor.source_transforms