+++ src/sphinx_codeautolink/extension/block.py
@@ -19,6 +19,7 @@ from .directive import ConcatMarker, PrefaceMarker, SkipMarker
 
 # list from https://pygments.org/docs/lexers/#pygments.lexers.python.PythonLexer
 BUILTIN_BLOCKS = {
+    "default": None,
     "python": None,
     "python3": None,
     "py": None,
@@ -104,7 +105,7 @@ else:
 
 
 class CodeBlockAnalyser(nodes.SparseNodeVisitor):
-    """Transform literal blocks of Python with links to reference documentation."""
+    """Transform blocks of Python with links to reference documentation."""
 
     def __init__(
         self,
@@ -113,6 +114,7 @@ class CodeBlockAnalyser(nodes.SparseNodeVisitor):
         global_preface: list[str],
         custom_blocks: dict[str, Callable[[str], str]],
         concat_default: bool,
+        default_highlight_lang: str | None,
         **kwargs,
     ) -> None:
         super().__init__(*args, **kwargs)
@@ -130,6 +132,7 @@ class CodeBlockAnalyser(nodes.SparseNodeVisitor):
         self.concat_section = False
         self.concat_sources = []
         self.skip = None
+        self.highlight_lang = default_highlight_lang
 
     def unknown_visit(self, node) -> None:
         """Handle and delete custom directives, ignore others."""
@@ -162,6 +165,10 @@ class CodeBlockAnalyser(nodes.SparseNodeVisitor):
     def unknown_departure(self, node) -> None:
         """Ignore unknown nodes."""
 
+    def visit_highlightlang(self, node) -> None:
+        """Set expected highlight language."""
+        self.highlight_lang = node["lang"]
+
     def visit_title(self, node) -> None:
         """Track section names and break concatenation and skipping."""
         self.title_stack.append(node.astext())
@@ -185,7 +192,7 @@ class CodeBlockAnalyser(nodes.SparseNodeVisitor):
 
     def visit_literal_block(self, node: nodes.literal_block):
         """Visit a generic literal block."""
-        return self.parse_source(node, node.get("language", None))
+        return self.parse_source(node, node.get("language", self.highlight_lang))
 
     def parse_source(
         self, node: nodes.literal_block | nodes.doctest_block, language: str | None